Externalization of Image and Mesh handling classes (Objects, Readers)   Moved out from samples.py to images.py and meshes.py   All are submodules of core module
